pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mwplu"
version = "0.3.0"
description = "This project offers an end-to-end solution for automated document management that handles extraction, analysis, and synthesis of document data. Using API calls including Mistral OCR and Gemini-pro, the system processes both text and images to generate synthesis according to a template from unstructured documents."
authors = [
    {name = "Grey Panda", email = "naviredelest@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "typer>=0.15.2,<0.16.0",
    "pyyaml>=6.0.2,<7.0.0",
    "loguru>=0.7.3,<0.8.0",
    "tqdm>=4.67.1,<4.68.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "mistralai>=1.6.0,<2.0.0",
    "google-genai>=1.9.0,<2.0.0",
    "pillow>=11.1.0,<12.0.0",
    "reportlab>=4.3.1,<5.0.0",
    "pypdf>=5.4.0,<6.0.0",
    "supabase>=2.15.0,<3.0.0",
    "svglib>=1.5.1,<2.0.0",
    "setuptools (>=80.9.0,<81.0.0)",
]